Using go generate to embed
Using `generate` to create a file from a template to create the variable
and comment for go embed. We do this as we are embeding Next.js which
makes use of files/dirs that start with underscores which go by default
treats as hidden files and does not embed them, so we just provide all
the paths.